requirements.txt
amqp==5.0.9
asgiref==3.4.1
autopep8==1.5.7
billiard==3.6.4.0
celery==5.2.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.28.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.6
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-redis-sessions==0.6.2
django-rest-swagger==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
kombu==5.2.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
prompt-toolkit==3.0.24
psycopg2==2.9.1
pycodestyle==2.7.0
PyJWT==2.3.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.3
PyYAML==5.4.1
redis==4.1.1
requests==2.26.0
simplejson==3.17.4
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
